Panning Start and Clock Spin
Made CameraMovement the sort of new GameManager Instance for changing game states. The game states are now the camera states instead because the UI changes based on the camera states. After loading, the camera pans to the tower and does a round click, then you can tap to spin and stop before camera starts following the player